requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.14
django-redis==5.2.0
djangorestframework==3.12.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.0.3
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
PyMySQL==1.0.2
python-memcached==1.59
pytz==2022.7
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
minio